requirements.txt
torch
transformers
tokenizers
numpy
pydantic
hydra-core
omegaconf
regex
requests
beautifulsoup4
unsloth